fix(directory): Style repo when root is home and we're in a subfolder
Suppose you want to track your dotfiles with git so `$HOME` is a git
repository and you have `repo_root_style` set because you want to
highlight the root of the repository you're currently in.

If you `cd /home/foo/.config` you would expect the prompt to highlight
the contracted `$HOME`, that is the `~` in `~/.config`.

What happens is that you instead get `~/.configfoo/.config` with `foo`
highlighted due to a conflict between the home contraction logic and
the repo root styler.

This patch makes things work in that scenario, but there are likely
other conflicts with other options in the module.

Signed-off-by: Emanuele Aina <em@nerd.ocracy.org>